+++ controlnet_ext/controlnet_ext.py
@@ -1,31 +1,37 @@
 from __future__ import annotations
 
 import importlib
+import re
 from functools import lru_cache
 from pathlib import Path
 
-from modules import sd_models, shared
+from modules import extensions, sd_models, shared
 from modules.paths import data_path, models_path, script_path
 
 ext_path = Path(data_path, "extensions")
 ext_builtin_path = Path(script_path, "extensions-builtin")
-is_in_builtin = False  # compatibility for vladmandic/automatic
 controlnet_exists = False
+controlnet_path = None
+cn_base_path = ""
 
-if ext_path.exists():
-    controlnet_exists = any(
-        p.name == "sd-webui-controlnet" for p in ext_path.iterdir() if p.is_dir()
-    )
+for extension in extensions.active():
+    if not extension.enabled:
+        continue
+    # For cases like sd-webui-controlnet-master
+    if "sd-webui-controlnet" in extension.name:
+        controlnet_exists = True
+        controlnet_path = Path(extension.path)
+        cn_base_path = ".".join(controlnet_path.parts[-2:])
+        break
 
-if not controlnet_exists and ext_builtin_path.exists():
-    controlnet_exists = any(
-        p.name == "sd-webui-controlnet"
-        for p in ext_builtin_path.iterdir()
-        if p.is_dir()
-    )
-
-    if controlnet_exists:
-        is_in_builtin = True
+cn_model_module = {
+    "inpaint": "inpaint_global_harmonious",
+    "scribble": "t2ia_sketch_pidi",
+    "lineart": "lineart_coarse",
+    "openpose": "openpose_full",
+    "tile": None,
+}
+cn_model_regex = re.compile("|".join(cn_model_module.keys()))
 
 
 class ControlNetExt:
@@ -35,45 +41,51 @@ class ControlNetExt:
         self.external_cn = None
 
     def init_controlnet(self):
-        if is_in_builtin:
-            import_path = "extensions-builtin.sd-webui-controlnet.scripts.external_code"
-        else:
-            import_path = "extensions.sd-webui-controlnet.scripts.external_code"
+        import_path = cn_base_path + ".scripts.external_code"
 
         self.external_cn = importlib.import_module(import_path, "external_code")
         self.cn_available = True
         models = self.external_cn.get_models()
-        self.cn_models.extend(m for m in models if "inpaint" in m)
+        self.cn_models.extend(m for m in models if cn_model_regex.search(m))
 
-    def _update_scripts_args(self, p, model: str, weight: float):
+    def update_scripts_args(
+        self, p, model: str, weight: float, guidance_start: float, guidance_end: float
+    ):
+        if (not self.cn_available) or model == "None":
+            return
+
+        module = None
+        for m, v in cn_model_module.items():
+            if m in model:
+                module = v
+                break
+
         cn_units = [
             self.external_cn.ControlNetUnit(
                 model=model,
                 weight=weight,
                 control_mode=self.external_cn.ControlMode.BALANCED,
-                module="inpaint_global_harmonious",
+                module=module,
+                guidance_start=guidance_start,
+                guidance_end=guidance_end,
                 pixel_perfect=True,
             )
         ]
 
         self.external_cn.update_cn_script_in_processing(p, cn_units)
 
-    def update_scripts_args(self, p, model: str, weight: float):
-        if self.cn_available and model != "None":
-            self._update_scripts_args(p, model, weight)
-
 
 def get_cn_model_dirs() -> list[Path]:
     cn_model_dir = Path(models_path, "ControlNet")
-    if is_in_builtin:
-        cn_model_dir_old = Path(ext_builtin_path, "sd-webui-controlnet", "models")
+    if controlnet_path is not None:
+        cn_model_dir_old = controlnet_path.joinpath("models")
     else:
-        cn_model_dir_old = Path(ext_path, "sd-webui-controlnet", "models")
+        cn_model_dir_old = None
     ext_dir1 = shared.opts.data.get("control_net_models_path", "")
     ext_dir2 = shared.opts.data.get("controlnet_dir", "")
 
-    dirs = [cn_model_dir, cn_model_dir_old]
-    for ext_dir in [ext_dir1, ext_dir2]:
+    dirs = [cn_model_dir]
+    for ext_dir in [cn_model_dir_old, ext_dir1, ext_dir2]:
         if ext_dir:
             dirs.append(Path(ext_dir))
 
@@ -81,7 +93,7 @@ def get_cn_model_dirs() -> list[Path]:
 
 
 @lru_cache
-def _get_cn_inpaint_models() -> list[str]:
+def _get_cn_models() -> list[str]:
     """
     Since we can't import ControlNet, we use a function that does something like
     controlnet's `list(global_state.cn_models_names.values())`.
@@ -98,7 +110,11 @@ def _get_cn_inpaint_models() -> list[str]:
             continue
 
         for p in base.rglob("*"):
-            if p.is_file() and p.suffix in cn_model_exts and "inpaint" in p.name:
+            if (
+                p.is_file()
+                and p.suffix in cn_model_exts
+                and cn_model_regex.search(p.name)
+            ):
                 if name_filter and name_filter not in p.name.lower():
                     continue
                 model_paths.append(p)
@@ -112,7 +128,7 @@ def _get_cn_inpaint_models() -> list[str]:
     return models
 
 
-def get_cn_inpaint_models() -> list[str]:
+def get_cn_models() -> list[str]:
     if controlnet_exists:
-        return _get_cn_inpaint_models()
+        return _get_cn_models()
     return []